refactor: do not used the 'shared' widgets model by default
Before, every element corresponded to a single widget and the
render function of the element was always executed in the context of the
render function where it was created.
This is ideal for bqplot, where you need the scales to reference
the same widget.
However, if you need to have a component that takes an element as
argument but sets some context, or catches exception, the
render function needs to be executed in the context of the
'deepest' element that created it.