feat: batch updates in event handlers
In 9502163 and 03a9b39 we implemented batch updates from on_<name>
change handlers. This caused a change handler that set multiple states
to be changed to not trigger a render multiple times.
This behaviour was not implemented for event handlers. This commit
implements this behaviour for event handlers as well.