fix: get_widget could fail with 'x not found in all known widgets'
This happened when we could re-render, and there was a function
component in the render tree that wouldn't need a re-execution.
It would then reconciliate with the old element data (also its arguments)
and the reference to the old element would map to get widget that
would be used in get_widget.
However, a use_effect/use_event would have a reference to a new element
and reacton would not be able to find the widget.
We now update the args and kwargs of the old element to the new
elements.